requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-ckeditor==6.4.0
django-cors-headers==3.11.0
django-js-asset==2.0.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
